Adding Internal Scope for Internal API call.
